[#3] feat: MemberErrorCode 추가
- SampleError 삭제
